requirements.txt
certifi==2019.6.16
dj-database-url==0.5.0
Django==2.2.3
django-bootstrap-modal-forms==1.4.2
django-crispy-forms==1.7.2
django-heroku==0.3.1
gunicorn==19.9.0
psycopg2==2.8.2
pytz==2019.1
sentry-sdk==0.10.1
sqlparse==0.3.0
urllib3==1.25.3
views==0.3
whitenoise==4.1.2
django-widget-tweaks==1.4.2